Fixed improper file structure and added docs comments
