--- src/src/main.rs
@@ -1,4 +1,0 @@
-use dsa::*;
-fn main() {
-    println!("Hello, world!");
-}
